Address review findings on the media split
Standards axis: the two media test files each carried a verbatim copy of
a_file/a_clip, and one test had its a_shallow_copy_pool call inlined
during the move. Both helpers plus the shadowed-copy pool move to
tests/mediapool.py, the way tests/cutfile.py and tests/otio.py already
hold shared fixtures, so the pair cannot drift on what a clip looks like.
resolve/pool.py takes get_logger("media.pool") - a child of the
operations logger, so handler config still catches it, but a live failure
says which half of the old module it came from.

Spec axis: resolve/media.py declares its own Clip/Pool = Any aliases
rather than importing them from .pool, matching every other module in the
package.

Refs #227
